requirements.txt
asgiref==3.2.9
astroid==2.4.1
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==3.0.7
django-appconf==1.0.4
django-bootstrap4==2.0.1
django-extensions==2.2.9
django-imagekit==4.0.2
django-phonenumber-field==4.0.0
django-random-image-from-folder==0.0.3
idna==2.9
importlib-metadata==1.6.1
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.0
phonenumbers==8.12.6
pickleshare==0.7.5
pilkit==2.0
Pillow==7.1.2
prompt-toolkit==3.0.5
Pygments==2.6.1
pylint==2.5.2
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.4
wrapt==1.12.1
zipp==3.1.0
